requirements.txt
Werkzeug==1.0.1
Flask==1.1.2
prometheus-client==0.8.0
requests
pyyaml
